Remove one scenario as we have one that covers the same + exception
